Creater Footer component and css for all pages
