+++ src/components/UI/Card.module.css
@@ -4,7 +4,6 @@
     padding: 1rem;
     box-shadow: 0 2px 8px #F68C3E;
     border-radius: 14px;
-    background-color: white;
     background-color: #e8ddcf;
 }
