+++ src/pages/Juges.js
@@ -7,8 +7,9 @@ import TextError from "../components/forms/TextError";
 import {useKeycloak} from "@react-keycloak/web";
 import keycloak from "../keycloak";
 import Card from "../components/UI/Card";
-import {useHistory} from "react-router-dom";
 import classes from "../components/UI/Card.module.css";
+import box from './Juges.module.css';
+import Footer from "../components/UI/Footer";
 
 const initialValues = {
     categoryId: '',
@@ -34,122 +35,114 @@ const Juges = () => {
         dispatch(fetchCategories())
     }, [])
 
-    //utilisation de useHistory() de react-router-dom comme trick pour ne plus voir les liens soulignés dans les boutons
-    const history = useHistory();
-
-    const goToResults = (path) => {
-        history.push(path);
-    }
 
     const categories = useSelector(state => state.categorie.data)
     const loadingState = useSelector(state => state.categorie.loading)
 
     const {keycloak} = useKeycloak();
 
-    const handleLogout = () => {
-        keycloak.logout({ redirectUri: 'http://localhost:3000' });
-    }
-    if(keycloak.authenticated) {
+
+    if (keycloak.authenticated) {
         return (
             <>
-            <Card>
-                <h2>Juges</h2>
-                {loadingState === 'Loading ...' && <h3>Chargement en cours</h3>}
+                <Card>
+                    {loadingState === 'Loading ...' && <h3>Chargement en cours</h3>}
 
-                {loadingState === 'fulfilled' &&
-                <Formik
-                    initialValues={initialValues}
+                    {loadingState === 'fulfilled' &&
+                    <Formik
+                        initialValues={initialValues}
 
-                    // envoi des données du formulaire vers l'api backend
-                    onSubmit={
-                        async (values, onSubmitProps) => {
-                            // creation du result correspondant au format de lápi backend
-                            const splittedName = values.competitor.split("+");
+                        // envoi des données du formulaire vers l'api backend
+                        onSubmit={
+                            async (values, onSubmitProps) => {
+                                // creation du result correspondant au format de lápi backend
+                                const splittedName = values.competitor.split("+");
 
-                            const result = {
-                                catId: values.categoryId,
-                                competitorFName: splittedName[0],
-                                competitorLName: splittedName[1],
-                                boulderDone: [values.boulderSuccess]
-                            }
+                                const result = {
+                                    catId: values.categoryId,
+                                    competitorFName: splittedName[0],
+                                    competitorLName: splittedName[1],
+                                    boulderDone: [values.boulderSuccess]
+                                }
 
 
-                            await axios.post('http://localhost:8081/juges/validation', result, axiosConfig)
-                                .then(response => {
-                                    setSucces(true);
-                                    setFailure(false);
-                                })
-                                .catch(e => {
-                                    setFailure(true);
-                                    setSucces(false);
-                                })
-                                .then(onSubmitProps.resetForm)
+                                await axios.post('http://localhost:8081/juges/validation', result, axiosConfig)
+                                    .then(response => {
+                                        setSucces(true);
+                                        setFailure(false);
+                                    })
+                                    .catch(e => {
+                                        setFailure(true);
+                                        setSucces(false);
+                                    })
+                                    .then(onSubmitProps.resetForm)
 
+                            }}
+                    >
+                        {formik => {
+                            return (
+                                <Form>
+
+                                    <div className={box.box}>
+                                        <label htmlFor='categoryId'>Sélectionner la catégorie du competiteur</label>
+                                        <Field className={classes.field} as='select'
+                                               id='categoryId' name='categoryId'>
+                                            <option>Selectionnez une categorie</option>
+                                            {categories.map((categorie) => {
+                                                return (
+                                                    <option key={categorie.id}
+                                                            value={categorie.id}>{categorie.categoryName}</option>
+                                                )
+                                            })}
+                                        </Field>
+                                        <ErrorMessage name='categoryId' component={TextError}/>
+                                    </div>
+
+                                    <div className={box.box}>
+                                        <label htmlFor='competitor'>Sélectionner le competiteur</label>
+                                        <Field className={classes.field} as='select' id='competitor' name='competitor'>
+                                            <option>Selectionnez un competiteur</option>
+                                            {/*regarde si 1er select rempli alors filtre les categorie pour récup les competiteurs de celle selectionnée*/}
+                                            {formik.values.categoryId !== "" && categories.filter(category => (
+                                                category.id === formik.values.categoryId))[0].competitors.map(comp => {
+                                                return (<option key={comp.firstname}
+                                                                value={`${comp.firstname}+${comp.lastname}`}>{comp.lastname} {comp.firstname}</option>)
+                                            })}
+                                        </Field>
+                                        <ErrorMessage name='categoryId' component={TextError}/>
+                                    </div>
+                                    <div className={box.box}>
+                                        <label htmlFor='boulderSuccess'>Selectionnez un bloc</label>
+                                        <Field className={classes.field} as='select' id='boulderSuccess'
+                                               name='boulderSuccess'>
+                                            <option>Selectionnez un bloc</option>
+                                            {/*regarde si 1er select rempli alors filtre les categorie pour récup les blocs de celle selectionnée*/}
+                                            {formik.values.categoryId !== "" && categories.filter(category => (
+                                                category.id === formik.values.categoryId))[0].boulders.map(bloc => {
+                                                return (<option key={bloc}
+                                                                value={bloc}>{bloc}</option>)
+                                            })}
+                                        </Field>
+                                        <ErrorMessage name='categoryId' component={TextError}/>
+                                    </div>
+                                    <div className={box.box}>
+                                        {/*bouton de reset du formulaire sansenvoi en cas d'echec du bloc*/}
+                                        <button className={box.judgeButton} style={{color: "white", backgroundColor: "red"}} type={"reset"}>Echec
+                                        </button>
+                                        <button className={box.judgeButton} style={{color: "white", backgroundColor: "green"}} type='submit'
+                                                disabled={!formik.isValid || formik.isSubmitting}>Réussite
+                                        </button>
+                                    </div>
+                                </Form>
+                            )
                         }}
-                >
-                    {formik => {
-                        return (
-                            <Form>
 
-                                <div>
-                                    <label htmlFor='categoryId'>Sélectionner la catégorie du competiteur</label>
-                                    <Field className={classes.field} as='select' id='categoryId' name='categoryId'>
-                                        <option>Selectionnez une categorie</option>
-                                        {categories.map((categorie) => {
-                                            return (
-                                                <option key={categorie.id}
-                                                        value={categorie.id}>{categorie.categoryName}</option>
-                                            )
-                                        })}
-                                    </Field>
-                                    <ErrorMessage name='categoryId' component={TextError}/>
-                                </div>
-
-                                <div>
-                                    <label htmlFor='competitor'>Sélectionner le competiteur</label>
-                                    <Field className={classes.field} as='select' id='competitor' name='competitor'>
-                                        <option>Selectionnez un competiteur</option>
-                                        {/*regarde si 1er select rempli alors filtre les categorie pour récup les competiteurs de celle selectionnée*/}
-                                        {formik.values.categoryId !== "" && categories.filter(category => (
-                                            category.id === formik.values.categoryId))[0].competitors.map(comp => {
-                                            return (<option key={comp.firstname}
-                                                            value={`${comp.firstname}+${comp.lastname}`}>{comp.lastname} {comp.firstname}</option>)
-                                        })}
-                                    </Field>
-                                    <ErrorMessage name='categoryId' component={TextError}/>
-                                </div>
-                                <div>
-                                    <label htmlFor='boulderSuccess'>Sélectionner le competiteur</label>
-                                    <Field className={classes.field} as='select' id='boulderSuccess' name='boulderSuccess'>
-                                        <option>Selectionnez un bloc</option>
-                                        {/*regarde si 1er select rempli alors filtre les categorie pour récup les blocs de celle selectionnée*/}
-                                        {formik.values.categoryId !== "" && categories.filter(category => (
-                                            category.id === formik.values.categoryId))[0].boulders.map(bloc => {
-                                            return (<option key={bloc}
-                                                            value={bloc}>{bloc}</option>)
-                                        })}
-                                    </Field>
-                                    <ErrorMessage name='categoryId' component={TextError}/>
-                                </div>
-                                {/*bouton de reset du formulaire sansenvoi en cas d'echec du bloc*/}
-                                <button style={{color: "white", backgroundColor: "red"}} type={"reset"}>Echec</button>
-                                <button style={{color: "white", backgroundColor: "green"}} type='submit'
-                                        disabled={!formik.isValid || formik.isSubmitting}>Réussite
-                                </button>
-                            </Form>
-                        )
-                    }}
-
-                </Formik>}
-                <button onClick={handleLogout}>Se déconnecter</button>
-            </Card>
-                <footer>
-                    <button onClick={handleLogout}>Se déconnecter</button>
-                    <button onClick={() => {goToResults('/welcome')}} type='button'>Accueil</button>
-                </footer>
+                    </Formik>}
+                </Card>
+                <Footer/>
             </>
         )
-    }else {
+    } else {
         return (
             <>
                 {keycloak.login()}
