+++ src/pages/Resultats.js
@@ -1,9 +1,11 @@
 import Card from "../components/UI/Card";
+import Footer from "../components/UI/Footer";
 
 const Resultats = () => {
     return (
         <Card>
             <h1>Page resultats</h1>
+            <Footer/>
         </Card>
     )
 }
